+++ ProjectEuler/Problems/801-900/801-810/Problem808.cs
@@ -30,7 +30,7 @@ internal sealed class Problem808 : IProblem
                 continue;
             }
 
-            var candidateInteger = (long)secondPrimeCandidate;
+            var candidateInteger = (int)secondPrimeCandidate;
 
             if (!this.primes.IsPrime(candidateInteger))
             {
